+++ core/mixins.py
@@ -1,5 +1,5 @@
 from django.core.exceptions import ObjectDoesNotExist
-from django.urls import reverse_lazy
+from core import constants
 
 
 class WagtailAdminExclusivePageMixin:
@@ -30,4 +30,8 @@ class EnableTourMixin:
 
 class AnonymousUserRequired:
     # used by core.wagtail_hooks.anonymous_user_required
-    anonymous_user_required_redirect_url = reverse_lazy('core:dashboard')
+    anonymous_user_required_redirect_url = constants.DASHBOARD_URL
+
+
+class AuthenticatedUserRequired:
+    authenticated_user_required_redirect_url = constants.LOGIN_URL