+++ CHANGELOG.md
@@ -25,9 +25,10 @@
 - MVP-579 - create export plan cms page
 - GP2-67 - fix backbutton upload logo
 - GP2-170 - business performance
-
+GP2-315 - route to market backend integration
 ### Implemented enhancements
 
+- GP2-279 - Dashboard page made wagtail configurable and added routing block
 - GP2-250 - Lesson - estimated time (FE)
 - GP2-252 - Lesson - objectives (FE)
 - GP2-79 - Lessons - title (FE)